scripts/requirements.txt
python-ldap
faker
python-dotenv
